local-ai/image-inference-server/requirements.txt
# Image inference server requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
diffusers==0.25.0
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0
accelerate==0.25.0
pillow==10.1.0
numpy==1.24.3
httpx==0.25.1
pydantic==2.5.0

